refactor: extract refresh helpers
